Add files via upload
This ULP can help to generate Pins in an Eagle library from a Pin list normally made 
available from chip manufacturer. 
Especially FPGA chips have a lot of I/O pins with long names. Rather than drawing a symbol 
for each bank, drawing single pin symbols and adding these to a library device allows using a 
part of the FPGA on different pages of the schematic by invoking the desired pins from a 
large chip on the sheet these pins are used. Invoked pins can be swapped by the Gateswap 
command in the schematic editor or can be deleted and invoked on a different schematic page 
later.